+++ Package.swift
@@ -10,9 +10,7 @@ let package = Package(
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
             name: "CoreDataPlus",
-            targets: ["CoreDataPlus"]),
-        .library(name: "CoreDataPlusSyntacticSugar",
-                 targets: ["CoreDataPlusSyntacticSugar"])
+            targets: ["CoreDataPlus"])
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -24,9 +22,6 @@ let package = Package(
         .target(
             name: "CoreDataPlus",
             dependencies: []),
-        .target(
-            name: "CoreDataPlusSyntacticSugar",
-            dependencies: ["CoreDataPlus"]),
         .testTarget(
             name: "CoreDataPlusTests",
             dependencies: ["CoreDataPlus"]),
