+++ Sources/CoreDataPlus/Documentation.docc/CoreDataPlus.md
@@ -1,0 +1,12 @@
+# ``CoreDataPlus``
+
+Overview and quick start are on the [GitHub README](https://github.com/spnkr/CoreDataPlus).
+
+> Warning: Beta documentation
+
+## Topics
+
+### <!--@START_MENU_TOKEN@-->Group<!--@END_MENU_TOKEN@-->
+
+- <!--@START_MENU_TOKEN@-->``Symbol``<!--@END_MENU_TOKEN@-->
+
