+++ Sources/CoreDataPlus/ManagedObjectCountable.swift
@@ -9,6 +9,12 @@ public protocol ManagedObjectCountable where Self: NSFetchRequestResult {
 }
 
 public extension ManagedObjectCountable {
+    
+    /// Gets a count of objects matching the passed Predicate.
+    /// - Parameters:
+    ///   - predicate: Any valid ```NSPredicate```
+    ///   - context: Any ``NSManagedObjectContext``
+    /// - Returns: `Int`
     static func countFor(_ predicate: NSPredicate?, context: NSManagedObjectContext) -> Int {
         
         let request = NSFetchRequest<Self>()
@@ -16,5 +22,16 @@ public extension ManagedObjectCountable {
         request.entity = entity()
 
         return (try? context.count(for: request)) ?? 0
+    }
+    
+    /// Gets a count of objects matching the passed Predicate.
+    /// - Parameters:
+    ///   - predicate: Any valid ```NSPredicate```
+    ///   - using: A ```ContextMode``` specifying a foreground or background context. You need to provide these contexts once, using ``CoreDataPlus/CoreDataPlus/setup(viewContext:backgroundContext:logHandler:)``. If you don't want to provide contexts, use ``CoreDataPlus/ManagedObjectCountable/countFor(_:context:)`` instead.
+    /// - Returns: Int
+    static func countFor(_ predicate: NSPredicate?, using: ContextMode = .foreground) -> Int {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        return countFor(predicate, context: context)
     }
 }
