+++ Sources/CoreDataPlus/ManagedObjectDeletable.swift
@@ -8,6 +8,12 @@ public protocol ManagedObjectDeletable {
 
 public extension ManagedObjectDeletable {
     
+    
+    /// jkdfjdfjns
+    /// - Parameters:
+    ///   - context: one
+    ///   - saveAfter: two
+    ///   - wrapInContextPerformBlock: three
     static func destroyAll(context: NSManagedObjectContext, saveAfter: Bool = true, wrapInContextPerformBlock: Bool = true) {
         if wrapInContextPerformBlock {
             context.perform {
@@ -33,5 +39,16 @@ public extension ManagedObjectDeletable {
         }
 
         try! context.save()
+    }
+    
+    /// abc
+    /// - Parameters:
+    ///   - using: ok
+    ///   - saveAfter: yes
+    ///   - wrapInContextPerformBlock: Nope
+    static func destroyAll(using: ContextMode = .foreground, saveAfter: Bool = true, wrapInContextPerformBlock: Bool = true) {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        destroyAll(context: context, saveAfter: saveAfter, wrapInContextPerformBlock: wrapInContextPerformBlock)
     }
 }
