+++ Sources/CoreDataPlus/ManagedObjectFindOrCreateBy.swift
@@ -60,6 +60,7 @@ public extension ManagedObjectFindOrCreateBy {
         return findButDoNotCreate(column: "id", value: id, context: context)
     }
 
+    
     static func findButDoNotCreate(column: String, value: Any, context: NSManagedObjectContext) -> Self? {
         
         let request = NSFetchRequest<Self>()
@@ -78,5 +79,33 @@ public extension ManagedObjectFindOrCreateBy {
         }
 
         return nil
+    }
+    
+    public static func findOrCreate(id: String, using: ContextMode = .foreground) -> Self {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        return findOrCreate(column: "id", value: id, context: context)
+    }
+    
+    /// Finds an instance of the NSManagedObject that has a column (property) matching the passed value. If it doesn't exist in the database, creates one, and returns it.
+    /// - Parameters:
+    ///   - column: Name of column that uniquely identifies a row (primary key).
+    ///   - value: Value of the primary key.
+    ///   - using: Optional. Specify `.foreground` for the main view context (operates on the main thread). Use `.background` to use a shared background context that automatically merges into the view context. Use `.custom(nsManagedObjectContext:)` to choose your own NSManagedObjectContext.
+    /// - Returns: An instance of the object
+    public static func findOrCreate(column: String, value: Any, using: ContextMode = .foreground) -> Self {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        return findOrCreate(column: column, value: value, context: context)
+    }
+    
+    static func findButDoNotCreate(id: String, using: ContextMode = .foreground) -> Self? {
+        return findButDoNotCreate(column: "id", value: id, using: using)
+    }
+    
+    static func findButDoNotCreate(column: String, value: Any, using: ContextMode = .foreground) -> Self? {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        return findButDoNotCreate(column: column, value: value, context: context)
     }
 }
