+++ Sources/CoreDataPlus/ManagedObjectSearchable.swift
@@ -28,4 +28,13 @@ public extension ManagedObjectSearchable {
         
         return (try? context.fetch(request)) ?? []
     }
+    
+    static func searchFor(_ predicate: NSPredicate?,
+                          sortBy: [NSSortDescriptor]? = nil,
+                          limit: Int? = nil,
+                          using: ContextMode) -> [Self] {
+        let context = contextModeToNSManagedObjectContext(using)
+        
+        return searchFor(predicate, sortBy: sortBy, limit: limit, context: context)
+    }
 }
