--- Sources/CoreDataPlusSyntacticSugar/ManagedObjectSearchable+Shortcuts.swift
@@ -1,14 +1,0 @@
-import CoreData
-import CoreDataPlus
-import Foundation
-
-public extension ManagedObjectSearchable {
-    static func searchFor(_ predicate: NSPredicate?,
-                          sortBy: [NSSortDescriptor]? = nil,
-                          limit: Int? = nil,
-                          using: ContextMode) -> [Self] {
-        let context = contextModeToNSManagedObjectContext(using)
-        
-        return searchFor(predicate, sortBy: sortBy, limit: limit, context: context)
-    }
-}
